Provide a debug adapter (#33)
This PR uses the new extension APIs to define a Swift-specific debug
adapter. We don't download anything, but just look for a suitable
`lldb-dap` binary on the user's system. On macOS, installing either
Xcode or the command-line tools brings along an lldb-dap that can read
Swift debuginfo and has pretty-printers set up, in contrast to CodeLLDB
which is not set up for Swift.

Before merging I'd like to add support for
[Swiftly](https://www.swift.org/blog/introducing-swiftly_10/) as well.

The name of the new adapter, for referring to in debug task definitions,
is `Swift`.