Bump to v0.4.3 (#36)
There was a bug in zed_extension_cli that prevented debug adapters in
extensions from being successfully registered. zed_extension_cli has
been fixed (https://github.com/zed-industries/extensions/pull/2927), but
we need to bump the version of this and the Swift extension so they'll
get rebuild and republished using the fixed CLI.

There are no functional changes in this release.